ADD) 미팅 방 참여, 퇴장 로직 구현

+++ gaebap/src/main/java/site/doggyyummy/gaebap/domain/meeting/dto/CreateMeetingResponseDTO.java
@@ -18,7 +18,7 @@ public class CreateMeetingResponseDTO {
     private String password;
 
     @JsonProperty("max_participant")
-    private Long maxParticipant;
+    private int maxParticipant;
 
     private String title;
